Correctly set fixtures order to respect dependencies
Some of the tests could not be run individually as some fixture was
missing or were in incorrect dependency order.

It was working just fine when running the full suite as these fixtures
are session scoped, so as long as the first test was ordered correct,
everything will work.

Strictly speaking, bitbake_path and bitbake_variables do not depend on
each other; this commit changes the order in some cases just for
consistency.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>