Package the shared test files
This packages the shared test files, namely

* fixtures.py
* common.py
* helpers.py

into a python package, and fixes everything needed in order for them to be
shared betwen meta-mender and mender-convert.

In addition, a parseropts.py file is created, which contains all the needed
configurations for pytest to load this as a separate package, with all the cli
flags, and configuration needed for the tests.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>