Remove "conversion" marker and logic to skip tests based on this
It seems to have been introduced by mistake during the Python3 upgrade
in November 2019.

Before this refactoring, there was no marker nor logic to skip any test
based on this; it was only used inside fixtures or specific tests to
modify their behaviour (for example, to decide on the path where the
image under test is expected to be).

See:
* https://github.com/mendersoftware/mender-image-tests/commit/a6cfbd1b09412784dcdd0dabaa67662129a8ff87
* https://github.com/mendersoftware/meta-mender/commit/a12742dc4aceeb6e560242590c5f42ce4c3868cc

Also, generally speaking, any test that is in this submodule is expected
to be run both on mender-convert and meta-mender (some configuration
exceptions may apply to the latter), but in any case a "conversion only"
test would need to live in mender-convert repository and not here.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>