test: Fix `--capture` option by removing Fabric.
We are using an extremely small subset of Fabric's features, which is
easily replicated with a small class calling ssh. The upside is that
it is much more compatible, and the `--capture` option does not need
to be set to `--capture=no` anymore. This is great for several
reasons:

* xdist doesn't support this mode.
* We can get per-test log output by using different `--capture`
  options (such as the default `--capture=fd`).
* The console isn't flooded with messages anymore.
* Only the log from failing tests is printed by default.

There is one downside, but this is possible to improve later: Output
is not printed while the command is running anymore, but only after it
has finished. This only has a practical effect if going back to the
`--capture=no` mode though, since the other modes anyway do not print
anything until the test has finished.

In addition, we need to change slightly how we run commands that need
to be in the background. Previously we could use
`multiprocessing.Process`, but this does not work anymore because
terminating the `Process` does not terminate the ssh sub-process, so
it sticks around even after killing it. We switch to providing a
`popen` flag instead, and then tests need to treat this as if they had
used `subprocess.Popen` directly. On the bright side, we can remove
`second_connection` which is not needed anymore since every call now
makes a new connection.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>