test: Make tests safe to run in parallel.
The vast majority consists of using temporary file names instead of
hard-coded ones. However, in a few cases we are using the worker index
as the unique factor, mostly for code simplicity. We also need to make
sure we use a unique port for the mock server.

Almost certainly I will have missed something, but we'll have to solve
that as we go.

Ticket: QA-426

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>