test: Fix race condition in `bitbake_variables` fixture.
The problem is that this particular fixture results in running Bitbake
in the original build directory, instead of the temporary one for each
worker. We are only reading variables, so semantically there is
nothing wrong with it, but Bitbake doesn't like it if two processes
try to do this at the same time. We cannot move it to the temporary
build directories either, because the setting up of those directories
depends on calling Bitbake in the original directory to get some key
variables.

So instead, make a lock file so that only one process can do this
particular Bitbake call at the same time. In practice this won't have
a great effect on the running time since this only affects one session
fixture.

Note in particular that we remove the default for the
`prepared_test_build` argument. It is still valid to use `None`, since
we need it to solve the problem above, but we want virtually all other
callers to pass this argument, so we want to avoid having `None` as a
default argument.

Ticket: QA-426

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>