test: Remove test for `mender.conf`, which is less relevant now.
Both conditions can be true; if Mender is installed, then the file
will be there, if it's not installed, then it won't. We don't track
this condition in the tests currently.

Since we now run package scripts, it's better to test this condition
in mender-dist-packages anyway, so just remove it from here.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>